+++ jason/src/peer/repo.rs
@@ -1,12 +1,12 @@
 use std::{collections::HashMap, rc::Rc};
 
-use anyhow::Result;
 use futures::channel::mpsc;
 use medea_client_api_proto::{IceServer, PeerId};
+use tracerr::Traced;
 
 use crate::media::MediaManager;
 
-use super::{PeerConnection, PeerEvent};
+use super::{PeerConnection, PeerError, PeerEvent};
 
 /// [`PeerConnection`] factory and repository.
 #[allow(clippy::module_name_repetitions)]
@@ -23,7 +23,7 @@ pub trait PeerRepository {
         events_sender: mpsc::UnboundedSender<PeerEvent>,
         enabled_audio: bool,
         enabled_video: bool,
-    ) -> Result<Rc<PeerConnection>>;
+    ) -> Result<Rc<PeerConnection>, Traced<PeerError>>;
 
     /// Returns [`PeerConnection`] stored in repository by its ID.
     fn get(&self, id: PeerId) -> Option<Rc<PeerConnection>>;
@@ -65,15 +65,18 @@ impl PeerRepository for Repository {
         peer_events_sender: mpsc::UnboundedSender<PeerEvent>,
         enabled_audio: bool,
         enabled_video: bool,
-    ) -> Result<Rc<PeerConnection>> {
-        let peer = Rc::new(PeerConnection::new(
-            id,
-            peer_events_sender,
-            ice_servers,
-            Rc::clone(&self.media_manager),
-            enabled_audio,
-            enabled_video,
-        )?);
+    ) -> Result<Rc<PeerConnection>, Traced<PeerError>> {
+        let peer = Rc::new(
+            PeerConnection::new(
+                id,
+                peer_events_sender,
+                ice_servers,
+                Rc::clone(&self.media_manager),
+                enabled_audio,
+                enabled_video,
+            )
+            .map_err(tracerr::map_from_and_wrap!())?,
+        );
         self.peers.insert(id, peer);
         Ok(self.peers.get(&id).cloned().unwrap())
     }