+++ jason/CHANGELOG.md
@@ -44,6 +44,9 @@ All user visible changes to this project will be documented in this file. This p
 - Signalling:
     - Emitting of RPC commands:
         - `AddPeerConnectionMetrics` with `IceConnectionState` ([#71](/../../pull/71)).
+- Error handling:
+    - Library API:
+        - `JasonError` as library error with trace information and underlying JS error if it is the cause ([#55])
 
 ### Fixed